requirements.txt
requests
beautifulsoup4
selenium
markdownify
python-dotenv
psycopg2-binary
tabulate
